Move old index tests to old_index_test.py
Move the old index_test.py tests (that don't integrate Elasticsearch) to
old_index_test.py so that I can write the new tests in index_test.py
without all that noise in the file and without  test class name collisions.
